refactor: JwtHandshakeInterceptor 위치 수정
